% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chefs}
\alias{chefs}
\title{chefs}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{chef}}{Chef name (full name)}
  \item{\code{nickname}}{Guy Fieri's nickname for the chef}
  \item{\code{handedness}}{Whether the chef is right-handed, left-handed, or
  ambidextrous}
  \item{\code{gender}}{Whether the chef is male, female, or non-binary}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Tournament_of_Champions_(TV_series)}
}
\usage{
data(chefs)
}
\description{
A dataset containing information on each Chef, which is the same for all
seasons
}
\examples{
library(tidyverse)
chefs \%>\%
  filter(gender == "female")
}
\keyword{datasets}
