% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{judges}
\alias{judges}
\title{judges}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season}}{Season number}
  \item{\code{episode}}{Episode number}
  \item{\code{round}}{Stage of the tournament: Round 1, Round 2, Quarterfinals, Semifinals, Final}
  \item{\code{judge}}{Name of guest judge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/https://en.wikipedia.org/wiki/Tournament_of_Champions_(TV_series)}
}
\usage{
data(judges)
}
\description{
A dataset containing information about who were the judges for each challenge/episode
}
\examples{
library(tidyverse)
judges \%>\%
  filter(judge == "Ming Tsai") \%>\%
  group_by(season,episode,round)
}
\keyword{datasets}
