% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results}
\alias{results}
\title{results}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season}}{Season number}
  \item{\code{episode}}{Episode number}
  \item{\code{round}}{Stage of the tournament: Round 1, Round 2, Quarterfinals, Semifinals, Final}
  \item{\code{challenge}}{Variable to help distinguish rounds within the same Coast & Round}
  \item{\code{coast}}{The coast depends on how many chefs start the competition. If there are 16 chefs, then the coasts are: East, West. If there are 32 chefs, then the coasts are: East A, East B, West A, West B}
  \item{\code{y}}{ Numeric value to help when creating the bracket}
  \item{\code{chef}}{ Name of chef}
  \item{\code{commentator}}{ Who presented their food to the judges: Simon Majumdar or Justin Warner? }
  \item{\code{order}}{ When did their food get presented to the judges: Presented 1st or Presented 2nd }
  \item{\code{score_taste}}{ Score that chef received for the taste of their dish: values of 0- 50}
  \item{\code{score_randomizer}}{ Score that chef received for how well they used the Randomizer: values of 0- 30}
  \item{\code{score_presentation}}{ Score that chef received for the presentation of their dish: values of 0- 20}
  \item{\code{total}}{ Total score that chef received: between 0 and 100 }
}
}
\source{
\url{https://en.wikipedia.org/wiki/https://en.wikipedia.org/wiki/Tournament_of_Champions_(TV_series)}
}
\usage{
data(results)
}
\description{
A dataset containing win and loss data for each chef in each episode
}
\examples{
library(tidyverse)
results \%>\%
  group_by(season) \%>\%
  filter(round == "Final") \%>\%
  select(chef,season,coast,total)
}
\keyword{datasets}
