% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{randomizer}
\alias{randomizer}
\title{randomizer}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season}}{Season number}
  \item{\code{episode}}{Episode number}
  \item{\code{round}}{Stage of the tournament: Round 1, Round 2, Quarterfinals, Semifinals, Final}
  \item{\code{challenge}}{Variable to help distinguish rounds within the same Coast & Round}
  \item{\code{coast}}{The coast depends on how many chefs start the competition. If there are 16 chefs, then the coasts are: East, West. If there are 32 chefs, then the coasts are: East A, East B, West A, West B}
  \item{\code{randomizer1}}{First wheel of randomizer}
  \item{\code{randomizer2}}{Second wheel of randomizer}
  \item{\code{randomizer3}}{Third wheel of randomizer}
  \item{\code{randomizer4}}{Fourth wheel of randomizer}
  \item{\code{time}}{Length of challenge. Unit is minutes}
}
}
\source{
\url{https://en.wikipedia.org/wiki/https://en.wikipedia.org/wiki/Tournament_of_Champions_(TV_series)}
}
\usage{
data(randomizer)
}
\description{
A dataset containing information about each challenge: protein, vegetables, equipment, style, time
}
\examples{
library(tidyverse)
randomizer \%>\%
   group_by(season,episode,challenge,coast) \%>\%
   filter(round == "Semi-final")
}
\keyword{datasets}
